Add save (create / update) functionality

+++ src/main/java/app/recipe/cookbook/recipe/dto/domain/InstructionDto.java
@@ -16,7 +16,7 @@ public class InstructionDto {
     private UUID id;
     private UUID recipeId;
 
-    private Short stepNumber;
+    private Integer stepNumber;
     private String content;
 
 }
